feat: Add Prelimiary Open Badge Credential support
Implemented naive and ugly. The existing types are in the way of making
this simple and sturdy. Existing VC types allow a lot of fallbacks to
versions that OBv3 doesn't support. The Sphereon VC types also often are
too lenient, allowing all sorts of structures and shapes that OBv3 never
can handle.

More precise and practical types aren't implemented yet. We rely on the
upstream :any:any: structures for now.

We push the request structure directly into the credential, without any
checks and mapping. This is a severe security issue and opens a lot of
use-cases where we'll be signing and issuing nonsense badges. See above
re: types for one way to solve this.

We focus on the simplest use case so that we can call it from our
Backpack wallet. Get that to work first.